requirements.txt
beautifulsoup4
aiohttp==3.6.2

async-timeout==3.0.1
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
discord.py==1.4.1
idna==2.9
idna-ssl==1.1.0
isort==5.5.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==4.7.5
pylint==2.6.0
requests==2.23.0
six==1.15.0
toml==0.10.1
typing-extensions==3.7.4.1
urllib3==1.25.8
websockets==8.1
wrapt==1.12.1
yarl==1.4.2

python-dotenv~=0.15.0